backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
alembic==1.12.1

# JSON & Data Processing
pandas>=2.2.0
numpy>=1.26.2

# Excel Processing
openpyxl==3.1.2

# Word Processing
python-docx==1.1.0

# PDF Processing
pdfplumber==0.10.3
camelot-py[cv]==0.11.0

# IFC/BIM Processing
ifcopenshell>=0.8.3

# OCR & Image Processing
pytesseract==0.3.10
opencv-python>=4.8.1.78
Pillow>=10.2.0

# NLP (optional - kann später installiert werden)
# spacy>=3.7.2  # Für Python 3.13: spacy>=3.8.0 benötigt

# AI/OpenAI
openai>=1.12.0

# Task Queue
celery==5.3.4
redis==5.0.1

# S3/Object Storage
boto3==1.29.7
minio==7.2.0

# PDF Generation
reportlab==4.0.7

# YAML Processing
pyyaml==6.0.1

# Authentication (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# PDF Plan Processing
pdf2image==1.16.3  # Optional für OCR-Fallback
PyMuPDF==1.23.8  # Für PDF-Bildextraktion

# Template Engine
jinja2==3.1.2

# Chart Generation
matplotlib==3.7.0
plotly==5.17.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
